+++ allauth/socialaccount/providers/openstreetmap/provider.py
@@ -1,5 +1,8 @@
 from allauth.socialaccount.providers.base import ProviderAccount
 from allauth.socialaccount.providers.oauth.provider import OAuthProvider
+from allauth.socialaccount.providers.openstreetmap.views import (
+    OpenStreetMapOAuthAdapter,
+)
 
 
 class OpenStreetMapAccount(ProviderAccount):
@@ -23,6 +26,7 @@ class OpenStreetMapProvider(OAuthProvider):
     id = "openstreetmap"
     name = "OpenStreetMap"
     account_class = OpenStreetMapAccount
+    oauth_adapter_class = OpenStreetMapOAuthAdapter
 
     def extract_uid(self, data):
         return data["id"]