+++ allauth/socialaccount/providers/trello/views.py
@@ -7,11 +7,9 @@ from allauth.socialaccount.providers.oauth.views import (
     OAuthLoginView,
 )
 
-from .provider import TrelloProvider
-
 
 class TrelloOAuthAdapter(OAuthAdapter):
-    provider_id = TrelloProvider.id
+    provider_id = "trello"
     request_token_url = "https://trello.com/1/OAuthGetRequestToken"
     authorize_url = "https://trello.com/1/OAuthAuthorizeToken"
     access_token_url = "https://trello.com/1/OAuthGetAccessToken"