+++ allauth/socialaccount/providers/twitter/provider.py
@@ -1,5 +1,6 @@
 from allauth.socialaccount.providers.base import AuthAction, ProviderAccount
 from allauth.socialaccount.providers.oauth.provider import OAuthProvider
+from allauth.socialaccount.providers.twitter.views import TwitterOAuthAdapter
 
 
 class TwitterAccount(ProviderAccount):
@@ -31,6 +32,7 @@ class TwitterProvider(OAuthProvider):
     id = "twitter"
     name = "Twitter"
     account_class = TwitterAccount
+    oauth_adapter_class = TwitterOAuthAdapter
 
     def get_auth_url(self, request, action):
         if action == AuthAction.REAUTHENTICATE: